services/python-api/requirements.txt
Flask==3.0.0
pandas==2.1.1
scikit-learn==1.3.1
google-cloud-storage==2.11.0
joblib==1.3.2
gunicorn==21.2.0